+++ CHANGELOG.md
@@ -81,6 +81,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Logger modifications
 
+- Cleanup batch add feature
+
 ### Removed
 
 - Removed seeders